🫩 Feat: add new transforms for RCNN and update dependencies
- Introduced `ToTensor` and `Normalize` classes in `transforms.py` for preprocessing images and labels.
- Updated `pyproject.toml` to include `pandas` and `scikit-image` as dependencies.
- Created a new `inference` module with a basic `main` function.
- Refactored the `training` module to streamline the training process and adjusted parameters for training.
- Updated `uv.lock` to reflect new dependencies and their versions.
